+++ packages/expo-brightness/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### 🎉 New features
 
+- Added `addBrightnessListener` to subscribe to brightness updates. Currently iOS only. ([#16851](https://github.com/expo/expo/pull/16851) by [@luoxuhai](https://github.com/luoxuhai))
+
 ### 🐛 Bug fixes
 
 ### 💡 Others